Firebase User型とtimetable関連の型エラーを修正
- settings.dartにfirebase_auth importを追加
- personal_time_table.dartの型キャストを追加
- timetable_repository.dartの型エラーを修正
- final var問題を修正

+++ lib/feature/timetable/personal_time_table.dart
@@ -106,7 +106,7 @@ class PersonalTimeTableScreen extends ConsumerWidget {
         Navigator.of(context).push(
           PageRouteBuilder(
             pageBuilder: (context, animation, secondaryAnimation) =>
-                PersonalSelectLessonScreen(term as int, week as int, period),
+                PersonalSelectLessonScreen(term as int, week as int, period as int),
             transitionsBuilder: fromRightAnimation,
           ),
         );
